Move tests/test_copy.py into tests/test_manage.py
